+++ Chauffeur/TextReaderExtensions.cs
@@ -1,0 +1,17 @@
+﻿using System.IO;
+using System.Threading.Tasks;
+
+namespace Chauffeur
+{
+    static class TextReaderExtensions
+    {
+        public static async Task<string> ReadLineWithDefaultAsync(this TextReader reader, string @default)
+        {
+            var input = await reader.ReadLineAsync();
+            if (string.IsNullOrEmpty(input))
+                return @default;
+
+            return input;
+        }
+    }
+}